[poc] distribute WebpackVisualizer.js
