+++ src/site/app.jsx
@@ -6,16 +6,14 @@ import readFile from '../shared/util/readFile';
 import formatSize from '../shared/util/formatSize';
 import {getAssetsData, getBundleDetails, ERROR_CHUNK_MODULES} from '../shared/util/stat-utils';
 import buildHierarchy from '../shared/buildHierarchy';
-
+import WebpackVisualizer from '../component/WebpackVisualizer';
 
 export default React.createClass({
     getInitialState() {
         return {
-            assets: [],
+            stats: null,
             needsUpload: true,
             dragging: false,
-            chartData: null,
-            selectedAssetIndex: 0
         };
     },
 
@@ -50,13 +48,9 @@ export default React.createClass({
 
     handleFileUpload(jsonText) {
         let stats = JSON.parse(jsonText);
-        let assets = getAssetsData(stats.assets, stats.chunks);
 
         this.setState({
-            assets,
-            chartData: buildHierarchy(stats.modules),
             needsUpload: false,
-            selectedAssetIndex: 0,
             stats
         });
     },
@@ -80,30 +74,6 @@ export default React.createClass({
         };
 
         request.send();
-    },
-
-    onAssetChange(ev) {
-        let selectedAssetIndex = Number(ev.target.value);
-        let modules, chartData, error;
-
-        if (selectedAssetIndex === 0) {
-            modules = this.state.stats.modules;
-        } else {
-            let asset = this.state.assets[selectedAssetIndex - 1];
-            modules = asset.chunk.modules;
-        }
-
-        if (modules) {
-            chartData = buildHierarchy(modules);
-        } else {
-            error = ERROR_CHUNK_MODULES;
-        }
-
-        this.setState({
-            chartData,
-            error,
-            selectedAssetIndex
-        });
     },
 
     renderUploadArea(uploadAreaClass) {
@@ -145,33 +115,14 @@ export default React.createClass({
             demoButton = <button onClick={this.loadDemo} className={demoClass} style={{marginTop: '0.5em'}}>Try a Demo</button>;
         }
 
-        if (this.state.stats){
-            bundleDetails = getBundleDetails({
-                assets: this.state.assets,
-                selectedAssetIndex: this.state.selectedAssetIndex
-            });
-        }
-
-        if (this.state.assets.length > 1) {
-            assetList = (
-                <div>
-                    <select onChange={this.onAssetChange} value={this.state.selectedAssetIndex}>
-                        <option value={0}>All Chunks</option>
-                        {this.state.assets.map((asset, i) => <option key={i} value={i + 1}>{asset.name} ({formatSize(asset.size)})</option>)}
-                    </select>
-                </div>
-            );
-        }
-
         return (
             <div>
                 <h1>Webpack Visualizer</h1>
 
-                {assetList}
-                {this.renderUploadArea(uploadAreaClass)}
-                {demoButton}
+                {this.state.needsUpload && this.renderUploadArea(uploadAreaClass)}
+                {this.state.needsUpload && demoButton}
 
-                <ChartWithDetails chartData={this.state.chartData} bundleDetails={bundleDetails} />
+                {this.state.stats && <WebpackVisualizer stats={this.state.stats} />}
 
                 {this.state.error && <div className="errorMessage">{this.state.error}</div>}
